+++ .idea/gradle.xml
@@ -34,6 +34,7 @@
             <option value="$PROJECT_DIR$/features/defaultApis" />
             <option value="$PROJECT_DIR$/features/easterEggs" />
             <option value="$PROJECT_DIR$/features/itemPlacements" />
+            <option value="$PROJECT_DIR$/features/navigation" />
             <option value="$PROJECT_DIR$/features/playground" />
             <option value="$PROJECT_DIR$/features/shaders" />
           </set>
