+++ core/ui/src/main/res/raw/cat.json
