+++ core/ui/src/main/res/raw/cat_hey.json
