+++ features/navigation/build.gradle.kts
@@ -5,3 +5,11 @@ plugins {
 android {
     namespace = "pro.jayeshseth.animations.navigation"
 }
+
+dependencies {
+    implementation(projects.core.ui)
+    implementation(projects.core.model)
+    implementation(projects.core.navigation)
+    implementation(libs.lottie)
+    implementation(libs.commonmodule)
+}
