+++ features/navigation/src/main/kotlin/pro/jayeshseth/animations/navigation/navigation/Routes.kt
@@ -1,0 +1,12 @@
+package pro.jayeshseth.animations.navigation.navigation
+
+import kotlinx.serialization.Serializable
+import pro.jayeshseth.animations.core.navigation.Route
+
+object NavigationRoutes {
+
+    @Serializable
+    data object Nav2GraphRoute : Route()
+
+
+}
