+++ features/navigation/src/main/kotlin/pro/jayeshseth/animations/navigation/screens/nav2/AniamtedNav.kt
@@ -1,4 +1,4 @@
-package pro.jayeshseth.animations.ui.screens
+package pro.jayeshseth.animations.navigation.screens.nav2
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
