+++ features/navigation/src/main/kotlin/pro/jayeshseth/animations/navigation/screens/nav2/animatedNav/ScaleNav.kt
@@ -1,4 +1,4 @@
-package pro.jayeshseth.animations.ui.animatedNav
+package pro.jayeshseth.animations.navigation.screens.nav2.animatedNav
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -10,7 +10,7 @@ import androidx.compose.ui.Modifier
 import com.airbnb.lottie.compose.LottieAnimation
 import com.airbnb.lottie.compose.LottieCompositionSpec
 import com.airbnb.lottie.compose.rememberLottieComposition
-import pro.jayeshseth.animations.R
+import pro.jayeshseth.animations.core.ui.media.AnimMedia
 
 @Composable
 fun ScaleNav() {
@@ -25,7 +25,7 @@ fun ScaleNav() {
 
 @Composable
 fun LottieCat1() {
-    val composition by rememberLottieComposition(LottieCompositionSpec.RawRes(R.raw.cat))
+    val composition by rememberLottieComposition(LottieCompositionSpec.RawRes(AnimMedia.catJson))
     LottieAnimation(
         composition,
     )
